html/semantics/scripting-1/the-script-element/bytes-module/sync-module.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Bytes modules</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
const t = async_test();
</script>
<script type="module" onerror="t.step(() => assert_unreached(event))">
import v from "./image.png" with { type: "bytes" };
t.step(() => {
  assert_true(v instanceof Uint8Array);
  const expected = [
    0x89,
    0x50,
    0x4e,
    0x47,
    0xd,
    0xa,
    0x1a,
    0xa,
    0x0,
    0x0,
    0x0,
    0xd,
    0x49,
    0x48,
    0x44,
    0x52,
    0x0,
    0x0,
    0x0,
    0x1,
    0x0,
    0x0,
    0x0,
    0x1,
    0x1,
    0x0,
    0x0,
    0x0,
    0x0,
    0x37,
    0x6e,
    0xf9,
    0x24,
    0x0,
    0x0,
    0x0,
    0xa,
    0x49,
    0x44,
    0x41,
    0x54,
    0x78,
    0x1,
    0x63,
    0x60,
    0x0,
    0x0,
    0x0,
    0x2,
    0x0,
    0x1,
    0x73,
    0x75,
    0x1,
    0x18,
    0x0,
    0x0,
    0x0,
    0x0,
    0x49,
    0x45,
    0x4e,
    0x44,
    0xae,
    0x42,
    0x60,
    0x82,
  ];
  assert_array_equals(Array.from(v), expected);
  t.done();
});
</script>
